requirements.txt
streamlit==1.40.2
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-dotenv==1.0.1
requests==2.32.3
httpx==0.28.1
ibm-watsonx-ai==1.1.26
openai==1.59.3
geopy==2.4.1
pandas==2.2.3
numpy==2.2.1
plotly==5.24.1
streamlit-folium==0.24.0
folium==0.19.4
